+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/modcompat/ModChecker.java
@@ -35,6 +35,7 @@ public class ModChecker {
 	public static boolean sodiumPresent = false;
 	public static boolean goodallPresent = false;
 	public static boolean backpackedPresent = false;
+	public static boolean projectileDamageAttributePresent = false;
 
 	/*
 	 * -- DO NOT TURN THE LAMBDAS INTO METHOD REFS. Method refs are not classloading safe. --
@@ -64,6 +65,9 @@ public class ModChecker {
 
 			modid = "backpacked";
 			loadupModCompat(modid, () -> new BackpackedCompat());
+
+			modid = "projectile_damage";
+			loadupModCompat(modid, () -> new ProjectileDamageAttributeCompat());
 		}
 		catch (Throwable e) {
 			printErrorToLogs("classloading " + modid + " and so, mod compat done afterwards broke");