+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-### **(V.6.7.10 Changes) (1.19.2 Minecraft)**
+### **(V.6.7.10 Changes) (1.19.4 Minecraft)**
 
 ##### Fluids:
 (Forge): Fixed Sugar Water not turning Lava into Obsidian. Forgot to do specific thingy with Forge's Fluid API system since they removed Lava's check for Water Tagged fluids.
+
+##### Items:
+Buffed Crystal Cannon to deal damage between 8 and 13 for each projectile and consumes 1 extra durability now.
+
+##### Mod Compat:
+Added compatibility with Projectile Damage Attribute mod with my Crystal Cannon. Should allow changing the weapon's damage and projectile speed through Projectile Damage Attribute mod.
 
 
 ### **(V.6.7.9 Changes) (1.19.4 Minecraft)**